diagnostic_api/requirements.txt
# ============================================================================
# STF AI Diagnosis Platform - Diagnostic API Dependencies
# Author: Li-Ta Hsu
# Date: January 2026
# ============================================================================
# Pinned versions for reproducibility (Phase 1)
# ============================================================================

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
alembic==1.13.1

# Vector Store
weaviate-client==4.4.1

# LLM Integration
openai==1.10.0  # OpenAI-compatible client for Ollama + OpenRouter
httpx==0.26.0

# Authentication
passlib[bcrypt]==1.7.4
bcrypt==4.2.1  # passlib 1.7.4 incompatible with bcrypt>=5
python-jose[cryptography]==3.3.0

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6

# Logging
structlog==24.1.0

# Validation
email-validator==2.1.0
phonenumbers==8.13.28

# PDF Processing
# TODO(3): Pin PyMuPDF to exact version for reproducibility (e.g., PyMuPDF==1.24.0)
PyMuPDF>=1.24.0

# CJK Text Processing (Chinese word segmentation for RAG chunker)
jieba>=0.42.1

# OCR (CJK + English text extraction from PDF images)
easyocr>=1.7.0

# Image Processing (used by vision service for resizing before LLM inference)
Pillow==12.1.0

# Testing (optional, for development)
pytest==7.4.4
pytest-asyncio==0.23.3
# TODO(4): Remove duplicate httpx - already listed under LLM Integration (line 25)
httpx==0.26.0
